--- src/main/java/pages/Marca.java
@@ -1,61 +1,0 @@
-package pages;
-
-import static org.openqa.selenium.support.PageFactory.initElements;
-import static utilities.DriverConfiguration.getDriver;
-
-import io.appium.java_client.pagefactory.AndroidFindBy;
-import io.appium.java_client.pagefactory.AppiumFieldDecorator;
-import java.time.Duration;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import utilities.Constants;
-import utilities.FrontEndOperation;
-
-public class Marca extends FrontEndOperation {
-  private WebDriver driver;
-
-  @FindBy(id = "ue-accept-notice-button")
-  WebElement acceptCookies;
-
-  @FindBy(partialLinkText = "Noticia que no existe")
-  @AndroidFindBy(
-      xpath =
-          "(//android.widget.LinearLayout[@resource-id=\"com.iphonedroid.marca:id/portadilla_container\"])[3]")
-  WebElement randomNotice;
-
-  @FindBy(id = "buttonYes")
-  WebElement ageButton;
-
-  @FindBy(className = "ue-c-cover-content__link")
-  @AndroidFindBy(id = "com.iphonedroid.marca:id/ue_cms_list_item_text_container")
-  WebElement noticia;
-
-  public Marca() {
-    driver = getDriver();
-    initElements(new AppiumFieldDecorator(driver, Duration.ofSeconds(Constants.LOW_TIMEOUT)), this);
-  }
-
-  public void acceptCookies() {
-    if (isVisible(acceptCookies)) {
-      acceptCookies.click();
-    }
-  }
-
-  public void goToNotice() {
-    randomNotice.click();
-  }
-
-  public void acceptAge() {
-    clickWhenVisible(ageButton);
-  }
-
-  public boolean isNoticeShow() {
-    return isVisible(randomNotice);
-  }
-
-  public void clickNoticia() {
-    clickWhenVisible(noticia);
-  }
-}
